+++ torchmetrics/classification/average_precision.py
@@ -126,3 +126,7 @@ class AveragePrecision(Metric):
         preds = dim_zero_cat(self.preds)
         target = dim_zero_cat(self.target)
         return _average_precision_compute(preds, target, self.num_classes, self.pos_label)
+
+    @property
+    def is_differentiable(self):
+        return False