+++ torchmetrics/classification/cohen_kappa.py
@@ -117,3 +117,12 @@ class CohenKappa(Metric):
         Computes cohen kappa score
         """
         return _cohen_kappa_compute(self.confmat, self.weights)
+
+    @property
+    def is_differentiable(self):
+        """
+        cohen kappa is not differentiable since the implementation
+        is based on calculating the confusion matrix which in general
+        is not differentiable
+        """
+        return False